split ImageSelectPanel completebutton handling into file ImageSelectPanelResut.js
